Update ST localizations through core api update
